beauty/requirements.txt
Django>=5.1.4,<7.0
djangorestframework>=3.14.0
django-import-export>=3.3.0
openpyxl>=3.1.0
Pillow>=10.0.0
django-filter>=23.0
django-simple-history>=3.4.0
flake8>=6.0.0

